haystack_dj/app_dj/search_indexes.py
from haystack import indexes
from models import Document


class DocumentIndex(indexes.SearchIndex, indexes.Indexable):
    text = indexes.CharField(document=True, use_template=True)

    def get_model(self):
        return Document
