requirements.txt
Django==1.7.4
django-haystack==2.3.1
elasticsearch==1.5.0
urllib3==1.10.4
